scripable-codelib/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"    # ← was "setuptools.backends.legacy:build"

[project]
name = "scriptable"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["requests", "pyyaml"]

[project.scripts]
scriptable = "scriptable.cli:main"

[tool.setuptools.packages.find]
where = ["."]